requirements.txt
PyQt6>=6.6.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0

# For Google Cloud Vision OCR integration in the Field Indexer (optional)
google-cloud-vision>=3.7.0

# For Gemini vision OCR integration (recommended for handwriting / free text)
google-genai>=1.0.0

